minor usability for demo web app
When clicking on the width/height fields, they now auto-highlight
the input field text. When someone clicks and then enters a number,
this will replace the number wholly instead of appending to whats
already there.

The field auto-updates the gilbert curve and, as of now,
only allows valid values (numeric, >0) and clamping to 1 if something invalid
is entered. Previously, if the user wanted to change the most
significant digit, they'd have to fight to manuver around the
digits so as not to have it auto-clamp to '1'. For example,
if the number there originally was '2', the user couldn't delete
'2', then put in 10, they'd have to put in '21', then manuver
the cursor in front to delete the '2', then append other numbers
after the '1'.

The usage is still a little clunky but the highlighting should mitigate
most of the standard usage.

One alternative is to provide a 'go' button that lets you update the
width/height fields without issue. I don't like this solution as
you can't see the curve update in real-time (by hitting the up/down
arrows in the number fields, say).

Another alternative is to do some type of 'debouncing' and only
update every 1s or so. I also don't like this as it's not real-time,
a little jarring and might be surprising to the user.

There might be something to be said for allowing empty or zero values,
doing a check and then updating when it's valid, but this gets
a little complicated and might have some weird edge case usuability issues
that I haven't thought of.

For now, I'm taking the progress-over-perfection approach and just adding
the text highlight solution.